# Cache set of undefined impossible case evars in evar_flags