--- a/engine/evarutil.mli
+++ b/engine/evarutil.mli
@@ -225,7 +225,7 @@ exception ClearDependencyError of Id.t * clear_dependency_error * GlobRef.t opti
     into an empty list. *)
 
 val restrict_evar : evar_map -> Evar.t -> Filter.t ->
-  ?src:Evar_kinds.t Loc.located -> constr list option -> evar_map * Evar.t
+  constr list option -> evar_map * Evar.t
 
 val clear_hyps_in_evi : env -> evar_map -> named_context_val -> types ->
   Id.Set.t -> evar_map * named_context_val * types