This avoids a fold_undefined in evarconv, significantly improving perf
on the cases with ma,y undefined evars in
https://github.com/coq/coq/issues/8233#issuecomment-1792508252

0 defined evars + 20k undefined evars: 1.6s -> 0.11s

60k defined evars + 20k undefined evars: 4.1s -> 0.75s

Close #8233