@JasonGross 2023-11-06
Nice!  Is it worth adding this example to coq-performance-tests (either one-off or in scaling)?