@SkySkimmer 2023-11-10
Cancelled the bench because without https://github.com/coq/coq/pull/18286 it will have a bunch of failures